+++ UniversityProject/Repository/Migrations/M004AddStudentTable.cs
@@ -9,13 +9,13 @@ public class M004AddStudentTable : AutoReversingMigration
     public override void Up()
     {
             Create.Table("Student")
-                .WithColumn("Id").AsInt32().PrimaryKey().Identity()
-                .WithColumn("PassportID").AsInt32().Unique().NotNullable().ForeignKey("Passport", "Id").OnDelete(Rule.Cascade)
-                .WithColumn("MilitaryID").AsInt16().ForeignKey("IdMilitary", "Id")
+                .WithColumn("Id").AsInt64().PrimaryKey().Identity()
+                .WithColumn("PassportID").AsInt64().Unique().NotNullable().ForeignKey("Passport", "Id").OnDelete(Rule.Cascade)
+                .WithColumn("MilitaryID").AsInt64().ForeignKey("IdMilitary", "Id")
                 .WithColumn("CriminalRecord").AsBoolean().NotNullable()
-                .WithColumn("CourseID").AsInt16().NotNullable()
-                .WithColumn("SkipHours").AsInt32().Nullable()
-                .WithColumn("CountOfExamsPassed").AsInt32().Nullable()
-                .WithColumn("CreditScores").AsInt32().Nullable();
+                .WithColumn("CourseID").AsInt64().NotNullable()
+                .WithColumn("SkipHours").AsInt64().Nullable()
+                .WithColumn("CountOfExamsPassed").AsInt64().Nullable()
+                .WithColumn("CreditScores").AsInt64().Nullable();
     }
 }
